Hot Module Replacement no longer crashes
HMR provided by webpack crashes, with the error: "Uncaught TypeError:
Cannot read property 'apply' of undefine" as shown in the following
issue on github:
https://github.com/webpack/webpack/issues/959

I'm not exactly sure what this issue is, as the cause isn't quite clear
to me from reading the bug report. But the latest code changes seemed to
have resolved the problem, as all of the root level modules that have
side-effects now properly support hot code reloading. This required
significant refactoring, and undoubtedly will require yet more.

The current issue is that whenever a change is made, a full reload is
required, and the only hint of issue is the following error message:
"[HMR] Error: Aborted because 125 is not accepted". Checking the file
that comes back from webpack indeed shows that 125 is empty. I'm
guessing this is some dependency such as freezer, rx, or what have you
that is causing the hot reload to fail.