Explore separate feature hierarchy from rendering
Inside EvenCrazierIdeas.js started exploring having a separate feature
and react component hierarchy. The feature hierarchy is much smaller,
and controls the application state, including the tie-in with RxJS.
Then, it hands off rendering to React, by way of renderer method, which
simply returns a stateless react component. The feature hierarchy and Rx
subscriptions and side-effects are built out with the build method,
though it's signature is still questionable given its responsibilities.

This is being considered as the go forward for how to structure the
application.