Use router, state per component, talk via kefir
Use the router in the application container, this means that each
component that the router uses now needs state, which has been
implemented as a freezer (state manager) per component. Component state
manager's listener API can then be used along with Kefir to expose
intentions. This allows for cross component communication. The root
container stores denormalized state for each of the components, and acts
as the central spot of registration of routing and state managers.

* switched from RxJS to Kefir as the latter is much smaller