Move component state into components
State is no longer completely centrally initialized. Instead, each
component defines the intial state for that it cares about, then
components that include it call the init method inside their models
(creating holes). This process rolls up recursively until the
initializes all models.

This also meant a centralized dispatcher is used, and instead of having
to dispatchFactory and actionCreator calls one dispatchFactory shares a
single dispatcher, and actions are namespaced by their creators. To
descend into the namespace, you ask for messages to be forwarded, in
which case their creator prefix appends the creator name. This entire
process is similar to nesting the models, as described above, except
we're nesting actions and their creation.