Set individual metrics in a meeting
Individual metrics can now be set inside a meeting. This surfaced a bug
in the way the history code was using freezer. Because history was being
bound to a specific instance of freezer state upon creation, the
histories were effectively branching. Instead, the creation of history
needs to fetch the freezer state each time, instead of being statically
bound to a specific state in the beginning.