Navigation works once more
Navigation broke after the the code was rearranged. Specifically, the
navigation broke when the Home was separated from AppContainer. Turns
out that the navigate trigger that Home was doing through the new
meeting button was actually only working by coincidence. When you
trigger a freezer event, such as was done with navigate, it will not
bubble up.

To remedy this there is no an effect property that can be passed around
everywhere, joy, that can trigger the relevant events at the top level.
New effects means new things being glomed onto this prop.